% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcomp21099R.R
\name{SurvFuncEsti}
\alias{SurvFuncEsti}
\title{Kaplan-Meier estimate for survival function}
\usage{
SurvFuncEsti(df)
}
\arguments{
\item{df}{the dataframe which consists of three columns:surviving time ,censoring or not(0,1),group}
}
\value{
a data frame consisists of eleven columns,adding Kaplan-Meier estimate for survival function at each time to the results in tableCalculation()
}
\description{
Kaplan-Meier estimate for survival function
}
\examples{
\dontrun{
SurvFuncEsti(aml1)
}
}
