% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcomp21099R.R
\name{survtest}
\alias{survtest}
\title{test if there is a difference between two survival curves.}
\usage{
survtest(df, B = 999, method = "FH", p = 1, q = 1, modified = TRUE)
}
\arguments{
\item{df}{the dataframe which consists of three columns:surviving time ,censoring or not(0,1),group.}

\item{B}{times for permutation.}

\item{method}{"FH" for Fleming and Harrington,"PP" for Peto & Peto method.}

\item{p}{a nonnegative parameter used in FM method.}

\item{q}{like p, controls the weight function S(t)^p(1-S(t))^q.}

\item{modified}{TRUE for modified Peto & Peto method,FALSE for not.}
}
\value{
a list with components:chisq_statistic for test statistic,p_val_chisq for p value based on big samples,p_val_pur for p value based on permutation test,method for the method you used.
}
\description{
test if there is a difference between two survival curves using Harrington and Fleming or Peto & Peto method,based on big sample theory and permutation test.
}
\examples{
\dontrun{
survtest(aml1,999,"FH",1,2)
survtest(aml1,999,"PP",modified=FALSE)
}
}
