% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcomp21099R.R
\name{tableCalculation}
\alias{tableCalculation}
\title{Calculations for some useful statistics}
\usage{
tableCalculation(df)
}
\arguments{
\item{df}{the dataframe which consists of three columns:surviving time ,censoring or not(0,1),group}
}
\value{
a dataframe consisists of ten columns,representing:"death occuring time","deathnumber at each time","total number at risk","expected death number in group1","expected death number in group2","number at risk in group1","number at risk in group2","variance of theoretical death number","death number in group1","death number in group2".
}
\description{
Calculate the statistic which is used to test difference between two survival curves
}
\examples{
\dontrun{
tableCalculation(aml1)
}
}
